+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `IconButton` supports `toggle` prop (uses `key` color when toggled and `aria-pressed`)
 - Improved test coverage / added image-snapshots
 - `ScheduleOutline` Icon artwork
+- `Tree` `noIndent` prop allows `Tree` to be indented at the same depth as adjacent `TreeItem`(s)
 
 ### Changed